[BE] refactor : 패키지 변경 및 참조 관계 수정
