+++ backend/reviewduck/src/main/java/com/reviewduck/review/dto/service/ReviewFormQuestionCreateDto.java
@@ -1,4 +1,4 @@
-package com.reviewduck.review.service;
+package com.reviewduck.review.dto.service;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
